backend updated

+++ migrations/20210129160858-create-appointment.js
@@ -19,7 +19,7 @@ module.exports = {
         }
       },
       slot_id: {
-        type: Sequelize.INTEGER.UNSIGNED,
+        type: Sequelize.BIGINT.UNSIGNED,
         references: {
           model: {
             tableName: 'slots',
@@ -46,4 +46,3 @@ module.exports = {
     await queryInterface.dropTable('Appointments');
   }
 };
-
